[REDEXPL] This is evil under amd64, too... so we disable it for now. [ROSHTTPD] UINT -> size_t [NCFTP] About time to clean this mess. Are we a Unix? No sir!
svn path=/trunk/; revision=67697

+++ rosapps/applications/net/roshttpd/common/list.cpp
@@ -34,7 +34,7 @@ CListNode::CListNode(PVOID element, CListNode *next, CListNode *prev)
 	Prev = prev;
 }
 
-void* CListNode::operator new(/*size_t*/ UINT size)
+void* CListNode::operator new(size_t size)
 {
     PVOID p;
     if (hHeap == NULL) {
